+++ src/components/queue-page/queue-page.tsx
@@ -6,68 +6,8 @@ import {Button} from "../ui/button/button";
 import {Circle} from "../ui/circle/circle";
 import {ElementStates} from "../../types/element-states";
 import {SHORT_DELAY_IN_MS} from "../../constants/delays";
-
-type TCircle = {
-  value: string;
-  state?: ElementStates;
-}
-
-interface IQueue<T> {
-  enqueue: (item: T) => void;
-  dequeue: () => void;
-  peak: () => T | null;
-  clear: () => void;
-  getArray: () => (T | null)[];
-  getHead: () => number;
-  getTail: () => number;
-  isEmpty: () => boolean;
-  isFull: () => boolean;
-}
-
-export class Queue<T> implements IQueue<T> {
-  private container: (T | null)[] = [];
-  private head = 0;
-  private tail = 0;
-  private readonly size: number = 0;
-  private length: number = 0;
-  constructor(size: number) {
-    this.size = size;
-    this.container = Array(size).fill(null);
-  }
-  enqueue = (item: T) => {
-    if (this.length >= this.size) {
-      throw new Error("Maximum length exceeded");
-    }
-    this.container[this.tail % this.size] = item;
-    this.tail++;
-    this.length++;
-  };
-  dequeue = () => {
-    if (this.isEmpty()) {
-      throw new Error("No elements in the queue");
-    }
-    this.container[this.head % this.size] = null;
-    this.head++;
-    this.length--;
-  };
-  peak = (): T | null => {
-    if (this.isEmpty()) {
-      throw new Error("No elements in the queue");
-    }
-    return this.container[this.head % this.size];
-  };
-  getHead = () => this.head % this.size;
-  getTail = () => (this.tail - 1) % this.size;
-  getArray = () => this.container;
-  isEmpty = () => this.length === 0;
-  isFull = () => this.length >= this.size;
-  clear = () => {
-    this.container = Array(this.size).fill(null);
-    this.head = 0;
-    this.tail = 0;
-    this.length = 0;
-  };
-}
+import {TCircle} from "../../types/circle";
+import {Queue} from "./queue";
 
 const QUEUE_SIZE = 7;
 
@@ -80,7 +20,7 @@ export const QueuePage: React.FC = () => {
   const onChange = useCallback((evt: ChangeEvent<HTMLInputElement>) => {
     evt.preventDefault();
     setInputString(evt.target.value);
-  }, [setInputString]);
+  }, []);
 
   useEffect(() => {
     setCircles([...Array(QUEUE_SIZE)].fill({ value: "", state: ElementStates.Default }))
@@ -102,22 +42,20 @@ export const QueuePage: React.FC = () => {
       input.dispatchEvent(e);
     }
     setInputString("");
-  }, [inputString, setCircles, queue]);
+  }, [inputString, queue]);
 
   const onDelete = useCallback(() => {
-    setCircles(circles.slice().map((item, index) => {
-      return index === queue.getHead() ? {...item, state: ElementStates.Changing} : item;
-    }));
+    setCircles(circles.slice().map((item, index) => index === queue.getHead() ? {...item, state: ElementStates.Changing} : item));
     setTimeout(() => {
       queue.dequeue();
       setCircles(queue.getArray().map(item => ({ value: item? item : "", state: ElementStates.Default })));
     }, SHORT_DELAY_IN_MS);
-  }, [setCircles, circles, queue]);
+  }, [circles, queue]);
 
   const reset = useCallback(() => {
     queue.clear();
     setCircles([...Array(QUEUE_SIZE)].fill({ value: "", state: ElementStates.Default }))
-  }, [queue, setCircles]);
+  }, [queue]);
 
   return (
     <SolutionLayout title="Очередь">
