+++ src/components/sorting-page/utils.ts
@@ -1,0 +1,5 @@
+export const swap = (arr: any[], firstIndex: number, secondIndex: number): void => {
+  const temp = arr[firstIndex];
+  arr[firstIndex] = arr[secondIndex];
+  arr[secondIndex] = temp;
+};
