+++ src/components/ui/solution-layout/solution-layout.module.css
@@ -22,7 +22,7 @@
   border-radius: 20px;
   margin: 31px 20px 28px;
   padding: 36px;
-  height: 693px;
+  min-height: 693px;
   box-sizing: border-box;
 }
