+++ src/types/circle.ts
@@ -1,0 +1,6 @@
+import {ElementStates} from "./element-states";
+
+export type TCircle = {
+  value: string;
+  state: ElementStates;
+}
